Bestelling afgewerkt

+++ Models/Domain/IBestellingRepository.cs
@@ -8,7 +8,7 @@ namespace WellStralerWebshop.Models.Domain
     public interface IBestellingRepository
     {
         public List<Bestelling> getBestellingen(KlantLogin klantLogin);
-        public Bestelling getBestellingById(int id, KlantLogin klantLogin);
+        public Bestelling getBestellingById(long id, KlantLogin klantLogin);
         public List<Bestelling> getBestellingenByFilter(KlantLogin klantLogin, string? productNaam, DateTime? vanDatum, DateTime? totDatum
             , string? leverAdres, byte? geleverd);
     }
